1st part fix server

+++ Certification_Backend/src/main/java/com/kwiky/backend/service/CanalDirectory.java
@@ -13,8 +13,7 @@ import com.kwiky.backend.model.Canal;
 public class CanalDirectory {
 	@Autowired
 	private CanalRepository canalRepository;
-
-
+	
 	// post Canal
 	public void addCanal(Canal newCanal) {
 		canalRepository.save(newCanal);
@@ -32,7 +31,8 @@ public class CanalDirectory {
 
 	// delete Canal by id
 	public void deleteCanal(Long id) {
-		canalRepository.deleteById(id);
+		if(canalRepository.existsById(id))
+			canalRepository.deleteById(id);
 	}
 
 	// Put Canal by id
@@ -47,6 +47,4 @@ public class CanalDirectory {
 	public List<Canal> searchByNameContains(String partialNameSearch) {
 		return canalRepository.findAllByNameContains(partialNameSearch);
 	}
-	
-	
 }
